requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx
python-dotenv
openpyxl